[LSS-9] add hooks
