--- src/class_components/Button/ButtonView.jsx
@@ -1,23 +1,0 @@
-import {Component} from 'react';
-import styles from './Button.module.scss';
-
-export default class ButtonView extends Component {
-
-    render() {
-        const {
-            text,
-            click,
-            children,
-            variant,
-            className
-        } = this.props;
-
-        return (
-            <button onClick={click} className={`${styles.button} ${styles[variant]}
-            ${className ? styles[className] : ''}`}>
-                {children}
-                <span className={styles['button-text']}>{text}</span>
-            </button>
-        )
-    }
-}
