+++ src/components/Gallery/GalleryLoader.jsx
@@ -1,0 +1,12 @@
+import {GalleryView} from "./GalleryView"
+
+export const GalleryLoader = () => {
+
+    const handlerImageLoader = () => {
+        console.log('Image is loaded');
+    }
+
+    return (
+        <GalleryView handlerImageLoader={handlerImageLoader}/>
+    )
+}
