+++ src/components/HOC/withLayout.jsx
@@ -1,0 +1,12 @@
+import {HeaderView} from "../Header/HeaderView";
+import {FooterView} from "../Footer/FooterView";
+
+export const withLayout = (WrappedComponent) => ({config, ...props}) => {
+    return (
+        <>
+            <HeaderView config={config}/>
+            <WrappedComponent {...props} />
+            <FooterView config={config}/>
+        </>
+    )
+}
