+++ src/components/Main/Main.module.scss
