--- src/function_components/Button/Button.module.scss
@@ -1,36 +1,0 @@
-@import "./src/styles/vars/vars.module.scss";
-
-.button-subscribe {
-  background: #8A5082;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  padding: 8px 24px;
-  border-radius: 32px;
-  width: 282px;
-  cursor: pointer;
-
-  &:hover {
-    background: $dark-hover-color;
-  }
-
-  .button-text {
-    color: $primary-text-color;
-    font-size: $font-size-md-3;
-    font-family: 'Calibri';
-    font-weight: $bold-weight;
-    padding-top: 3px;
-    margin-left: 10px;
-  }
-}
-
-
-
-
-
-
-
-
-
-
