--- src/function_components/Card/Card.module.scss
@@ -1,121 +1,0 @@
-@import "./src/styles/vars/vars.module.scss";
-
-.card {
-  display: flex;
-  flex-direction: column;
-  padding: 24px;
-  align-items: center;
-  position: relative;
-  width: 330px;
-  z-index: 1;
-  margin-bottom: 63px;
-
-  .custom-card {
-    z-index: -1;
-    &:before {
-      background: url('../../assets/svg/card-border.svg');
-      content: "";
-      background-repeat: no-repeat;
-      background-size: contain;
-      position: absolute;
-      width: 332px;
-      height: 520px;
-      top: -11px;
-      right: 0;
-    }
-
-    &:after {
-      background: url('../../assets/svg/card-bg.svg');
-      background-repeat: no-repeat;
-      content: '';
-      width: 330px;
-      height: 524px;
-      position: absolute;
-      z-index: -1;
-      background-size: contain;
-      top: -13px;
-      left: 0;
-    }
-
-  }
-
-
-  .product-image {
-    position: relative;
-    margin-top: -14px;
-
-    &:after {
-      background: url('../../assets/svg/bg.svg');
-      content: '';
-      width: 100%;
-      height: 100%;
-      position: absolute;
-      background-repeat: no-repeat;
-      top: 6px;
-      z-index: -1;
-      right: 3px;
-    }
-
-    .image {
-      width: 100%;
-    }
-  }
-
-  .product-info {
-    display: flex;
-    flex-direction: column;
-    margin-top: 14px;
-    margin-bottom: 24px;
-    width: 267px;
-
-    .title-container {
-      display: flex;
-      flex-direction: column;
-
-      .title {
-        font-weight: $bold-weight;
-        font-size: $font-size-lg;
-        color: $secondary-text-color;
-      }
-    }
-
-
-    .description {
-      font-size: $font-size-md;
-      color: $text-color;
-      font-weight: $regular-weight;
-      line-height: 34px;
-    }
-
-    .info {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-between;
-      align-items: center;
-      margin-top: 17px;
-
-      .price {
-        font-weight: $bold-weight;
-        font-size: $font-size-lg;
-        color: $secondary-text-color;
-      }
-
-      .weight {
-        font-weight: 300;
-        font-size: 20px;
-        color: $font-size-md-2;
-      }
-    }
-  }
-}
-
-
-
-
-
-
-
-
-
-
-
