--- src/function_components/Footer/Footer.module.scss
@@ -1,80 +1,0 @@
-@import './src/styles/vars/vars.module.scss';
-
-
-.footer {
-  background-color: $background;
-  padding-top: 24px;
-  padding-bottom: 23px;
-
-  .container {
-    max-width: 1160px;
-    margin: 0 auto;
-
-    .footer-container {
-      display: flex;
-      flex-direction: row;
-      align-items: center;
-      justify-content: space-between;
-      width: 100%;
-
-
-      .footer-info {
-        display: flex;
-        flex-direction: column;
-        color: $primary-text-color;
-        font-family: 'Calibri';
-
-        nav {
-          ul {
-            display: flex;
-            flex-direction: row;
-            margin-bottom: 16px;
-            justify-content: space-between;
-            font-family: 'Calibri';
-
-            li {
-              a {
-                font-weight: $bold-weight;
-                color: $primary-text-color;
-                font-size: $font-size-md;
-                letter-spacing: 0.5px;
-                padding: 0 21px;
-
-                &:hover {
-                  color: $hover-color;
-                }
-              }
-
-            }
-          }
-        }
-
-        .copyright {
-          text-align: center;
-          font-size: 14px;
-        }
-      }
-
-      .creator {
-        display: flex;
-        flex-direction: column;
-
-        ul {
-          font-family: 'Calibri';
-
-          li {
-            font-size: $font-size-md-2;
-            color: $primary-text-color;
-
-            a.creator-link {
-              font-size: $font-size-md-2;
-              color: $primary-text-color;
-            }
-          }
-        }
-      }
-    }
-  }
-}
-
-
