--- src/function_components/Main/Main.module.scss
@@ -1,91 +1,0 @@
-@import './src/styles/vars/vars.module.scss';
-
-.container {
-  max-width: 1160px;
-  margin: 0 auto;
-
-  .jumbotron {
-    font-family: 'Calibri';
-    margin-bottom: 112px;
-    position: relative;
-
-    .subtitle {
-      margin-top: 19px;
-      margin-bottom: 32px;
-      display: flex;
-      flex-direction: column;
-      width: 314px;
-
-
-      .description {
-        color: $secondary-text-color;
-        font-weight: $regular-weight;
-        font-size: $font-size-lg;
-        margin-top: 43px;
-        margin-bottom: 32px;
-      }
-    }
-  }
-
-
-  .strawberry {
-    background-image: url('../../assets/img/banner/strawberry-1.png');
-    width: 150px;
-    height: 168px;
-    background-repeat: no-repeat;
-    top: 127px;
-    left: 425px;
-    position: absolute;
-  }
-
-  .second-strawberry {
-    background-image: url('../../assets/img/banner/strawberry-2.png');
-    position: absolute;
-    background-repeat: no-repeat;
-    width: 150px;
-    height: 168px;
-    top: 110px;
-    right: 80px;
-    z-index: -1;
-  }
-
-
-  .background-zephir {
-    background-image: url('../../assets/img/banner/Main-zephir.png');
-    position: absolute;
-    right: 50px;
-    top: 0px;
-    width: 442px;
-    height: 437px;
-  }
-
-  .background {
-    background-image: url('../../assets/img/banner/cloud-2.png');
-    position: absolute;
-    right: 0;
-    background-repeat: no-repeat;
-    top: 45px;
-    height: 843px;
-    background-size: cover;
-    width: 65%;
-  }
-
-  section#cards {
-    font-family: 'Calibri';
-    margin-top: 52px;
-
-    .cards-container {
-      display: flex;
-      justify-content: center;
-    }
-
-    .wrapper {
-      display: flex;
-      flex-wrap: wrap;
-      width: 100%;
-      justify-content: space-between;
-    }
-  }
-}
-
-
