+++ src/pages/OrderTable/Modal/ModalView.jsx
@@ -1,6 +1,6 @@
 import {Component} from "react";
 import styles from './ModalView.module.scss'
-import ButtonView from "../../../class_components/Button/ButtonView";
+import {ButtonView} from "../../../components/Button/ButtonView";
 import {BUTTON} from "../../../constants/constants";
 
 export class ModalView extends Component {
