+++ src/pages/Todo/Todo.jsx
@@ -1,6 +1,6 @@
 import {Component} from "react";
 import TodoView from "./TodoView";
-import withLayout from "../../class_components/HOC/withLayout";
+import {withLayout} from "../../components/HOC/withLayout";
 import {THEMES, LABEL} from "../../constants/constants";
 import React from "react";
